+++ apps/dyn_api/apps.py
@@ -7,4 +7,4 @@ from django.apps import AppConfig
 
 class DynApiConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'dyn_api'
+    name = 'apps.dyn_api'