+++ apps/dyn_dt/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 class DynDtConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'dyn_dt'
+    name = 'apps.dyn_dt'