+++ apps/pages/urls.py
@@ -4,5 +4,5 @@ from django.contrib.auth import views as auth_views
 from . import views
 
 urlpatterns = [
-  path(''       , views.index,  name='index'),
+  path('', views.index,  name='index'),
 ]