+++ build.sh
@@ -13,4 +13,3 @@ python manage.py collectstatic --no-input
 python manage.py makemigrations
 python manage.py migrate
 
-python manage.py generate-api -f